Removed Material, mesh now has textures.
