+++ StortSpelProjekt/Engine/src/ECS/SceneManager.cpp
@@ -9,7 +9,6 @@
 #include "../Renderer/CommandInterface.h"
 #include "../Renderer/ShaderResourceView.h"
 #include "../Renderer/ConstantBufferView.h"
-#include "../Renderer/Material.h"
 #include "../Renderer/ShadowInfo.h"
 #include "../Renderer/ViewPool.h"
 #include "../Renderer/Texture.h"
